Allow multiple inputs and validate input data
